requirements-main.txt
pyperclip~=1.8.2
requests~=2.31.0
customtkinter~=5.2.0
flask~=3.0.0
Pillow~=10.1.0
qrcode~=7.4.2
